+++ src/ForgotPassword/index.js
@@ -1,14 +1,14 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import './index.css';
-import { usePasswordReset } from "../hooks/usePasswordReset";
+import { useRequestPasswordReset } from "../hooks/useRequestPassword";
 
 
 
 const ForgotPassword = () => {
   const [email, setEmail] = useState("");
   const [message, setMessage] = useState("");
-  const { sendResetRequest, loading, error } = usePasswordReset();
+  const { sendResetRequest, loading, error } = useRequestPasswordReset();
   const navigate = useNavigate();
 
   const handleSubmit = async (e) => {
